cmake: Fix GetGitRevisionDescription cmake module for MSys
When running in an msys environment and in a git worktree directory,
the git revision is not found.

This patch converts the git_worktree_dir variable to the windows format,
by adding the windows prefix of the msys root dir, so that the HEAD file
of the worktree can be found and used by the module.